requirements.txt
annotated-types==0.7.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
behave==1.2.6
blinker==1.9.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.3.0
execnet==2.1.2
fasteners==0.20
filelock==3.20.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.21.3
mdurl==0.1.2
mycdp==1.3.0
outcome==1.3.0.post0
packaging==25.0
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.6
pdbp==1.8.1
platformdirs==4.5.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pynose==1.5.5
pyotp==2.9.0
PySocks==1.7.1
pytest==9.0.1
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
sbvirtualdisplay==1.4.0
selenium==4.38.0
seleniumbase==4.44.12
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
tabcompleter==1.4.0
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.3.1
